Clean up monster classes
