+++ results.csv
@@ -1,10 +1,10 @@
 ,Jaw Worm,Fat Gremlin,Gremlin Wizard,Red Slaver,Pointy,Centurion,Mad Gremlin,Blue Slaver,Cultist
-Jaw Worm,617,1000,982,123,563,0,1000,367,404
+Jaw Worm,616,1000,981,129,580,0,1000,387,388
 Fat Gremlin,0,1000,0,0,0,0,0,0,0
-Gremlin Wizard,43,1000,1000,0,0,0,1000,0,0
+Gremlin Wizard,45,1000,1000,0,0,0,1000,0,0
 Red Slaver,1000,1000,1000,1000,1000,0,1000,1000,1000
-Pointy,689,1000,1000,0,1000,0,1000,0,1000
-Centurion,1000,1000,1000,1000,1000,820,1000,1000,1000
+Pointy,692,1000,1000,0,1000,0,1000,0,1000
+Centurion,1000,1000,1000,1000,1000,814,1000,1000,1000
 Mad Gremlin,0,1000,0,0,0,0,1000,0,0
-Blue Slaver,845,1000,1000,230,1000,0,1000,934,1000
-Cultist,841,1000,1000,0,1000,0,1000,587,1000
+Blue Slaver,859,1000,1000,219,1000,0,1000,912,1000
+Cultist,841,1000,1000,0,1000,0,1000,596,1000
